+++ src/media/data/media/contents/movie/internal.py
@@ -44,6 +44,7 @@ class Movie(AbstractContent):
         super().__init__()
         self.technical = None
         self.crew = None
+        self.s_index = None
         self._process(in_element)
 
     def build_index_object(self):
@@ -68,6 +69,10 @@ class Movie(AbstractContent):
                 self.crew = Crew(child)
         self._post_load_process()
 
+    def _post_load_process(self):
+        super()._post_load_process()
+        self.s_index = MovieIndexEntry(self)
+
     def __hash__(self):
         return hash(self.unique_key)
 
@@ -76,9 +81,6 @@ class Movie(AbstractContent):
 
     def __gt__(self, other):
         return self.unique_key > other.unique_key
-
-    def __eq__(self, other):
-        return self.unique_key == other.unique_key
 
 
 class MovieIndexEntry(ContentIndex):