+++ src/media/tools/audio/albums/list.py
@@ -22,26 +22,59 @@
 # SOFTWARE.
 #
 
-"""
-Simple command to output details on every movie.
-"""
+'''
+List all albums (or a random sample of albums),
+one entry per line.
+'''
 
 # pylint: disable=R0801
 # pylint: disable=too-few-public-methods
 
-# brief movie report
+import os
 import argparse
-import media.fmt.text.movie
+import media.fmt.text.audio.album
 from media.fileops.filenames import FilenameMatches
 from media.fileops.repo import Repo
-from media.tools.movies.common import (Controller,
-                                       MovieReport)
+from media.generic.sorting.content.audio import AlbumOrganizer
+from media.generic.sorting.batch import Batch
 
 
-class ShowController(Controller):
+CliGroupingOptions = {
+        'none': AlbumOrganizer.G_NONE,
+        'alphabetical': AlbumOrganizer.G_ANY_ALPHA,
+        'artist': AlbumOrganizer.G_ARTIST,
+        'decade': AlbumOrganizer.G_ANY_DECADE,
+        'genre': AlbumOrganizer.G_ANY_GENRE,
+}
+
+
+CliSortOptions = {
+        'title': Batch.S_TITLE,
+        'year': Batch.S_YEAR,
+        'runtime': Batch.S_RUNTIME,
+        'artist': Batch.S_ARTIST,
+}
+
+
+class Controller():
     '''
-    Subclass of default Controller class.
+    The Controller class manages aspects of the command
+    execution.  It processes all of the CLI arguments,
+    and tests file paths to make sure they exist.
     '''
+    def __init__(self):
+        self.args = None
+        self.mediapath = None
+
+    def setup(self):
+        '''
+        Initialize the command line parser, and make sure
+        the mediapath exists.
+        '''
+        parser = self._setup_parser()
+        self.args = parser.parse_args()
+        self._determine_path()
+
     def _setup_parser(self):
         '''
         Set up the parser object.
@@ -52,35 +85,70 @@ class ShowController(Controller):
                             help='print X random entries')
         parser.add_argument('--group',
                             choices=['none', 'alphabetical',
-                                     'decade',
+                                     'artist', 'decade',
                                      'genre'],
                             help='Album grouping',
                             default='none')
         parser.add_argument('--sort',
-                            choices=['title', 'year', 'runtime'],
+                            choices=['title', 'artist', 'year', 'runtime'],
                             help='Album sorting',
                             default='title')
-        parser.add_argument('--pagebreaks',
-                            action=argparse.BooleanOptionalAction,
-                            help='Page breaks on output')
         parser.add_argument('--stats',
                             action=argparse.BooleanOptionalAction,
                             help='Report statistics')
         return parser
 
+    def _determine_path(self):
+        '''
+        Verify that the media path with the XML files exists.
+        '''
+        mediapath = self.args.mediapath
+        if not mediapath:
+            if 'MEDIAPATH' in os.environ:
+                mediapath = os.environ['MEDIAPATH']
+        self.mediapath = mediapath
 
-class MovieShowReport(MovieReport):
+
+class AlbumList():
     '''
-    Subclass of MovieReport class.
+    Class that does all the listing work.
+
+    Should be able to take a list of albums,
+    and present that same list a multitide of
+    ways.
+
+    Should also be directly callable from
+    an interactive Python prompt.
     '''
     def __init__(self):
-        super().__init__()
-        self.pagebreaks = False
+        self.organizer = None
+        self.albums = []
+        self.sample = 0
+        self.group = AlbumOrganizer.G_ANY_ALPHA
+        self.asort = Batch.S_TITLE
+        self.stats = False
+
+    def set_albums(self, in_albums, in_controller=None):
+        '''
+        Pass all the Album content objects that were
+        found in the repository.
+        '''
+        if len(in_albums) > 0:
+            self.albums = in_albums
+            self.organizer = AlbumOrganizer(in_albums)
+            if in_controller:
+                self.params_from_controller(in_controller)
 
     def params_from_controller(self, in_controller):
-        super().params_from_controller(in_controller)
-        if in_controller.args.pagebreaks:
-            self.pagebreaks = True
+        '''
+        Get the parameters from the controller object.
+        '''
+        self.group = CliGroupingOptions[in_controller.args.group]
+        self.asort = CliSortOptions[in_controller.args.sort]
+        if in_controller.args.random:
+            self.sample = in_controller.args.random
+        if in_controller.args.stats:
+            self.stats = True
 
     def report(self, grouping=None, sorting=None, sample=0, stats=False):
         '''
@@ -96,7 +164,7 @@ class MovieShowReport(MovieReport):
         if stats:
             self.stats = stats
         out = ''
-        # out += media.fmt.text.movie.OneLiner.header_fields()
+        out += media.fmt.text.audio.album.OneLiner.header_fields()
         self.organizer.set_grouping(self.group)
         if self.sample > 0:
             batches = self.organizer.create_batches(self.group, self.sample)
@@ -108,22 +176,21 @@ class MovieShowReport(MovieReport):
             for batch_i in sorted(batches):
                 out += f"  -- {batch_i.header} ({len(batch_i.entries)}) --\n"
                 out += self._out_batch(batch_i, self.asort)
-            out = out[:-1]
-        # out += media.fmt.text.movie.OneLiner.header_line()
+        out = out[:-1]
+        out += media.fmt.text.audio.album.OneLiner.header_line()
         if self.stats:
             out += self._stats()
         return out
 
     def _stats(self):
         '''
-        Output the basic stats on the number of movies,
-        and the number of movies in the sample set
+        Output the basic stats on the number of albums,
+        and the number of albums in the sample set
         (if one is defined)
         '''
         all_c = len(self.organizer.entries)
         wrk_c = len(self.organizer.working)
-        out = "\n    ---- Movie Statistics ----\n\n"
-        out += f"  {'Movie count':12s} : {all_c:5d}\n"
+        out = f"  {'Album count':12s} : {all_c:5d}\n"
         if wrk_c < all_c:
             wrk_p = float(wrk_c) / all_c * 100
             out += f"  {'Sample count':12s} : {wrk_c:5d} ({wrk_p:5.2f}%)\n"
@@ -135,11 +202,9 @@ class MovieShowReport(MovieReport):
         '''
         out = ''
         order_list = batch.index_by(sort_field)
-        for movie in order_list:
-            out += str(media.fmt.text.movie.Brief(movie.movie))
-            out += "\n"
-            if self.pagebreaks:
-                out += chr(12)
+        for album in order_list:
+            out += str(media.fmt.text.audio.album.OneLiner(album.album))
+        out += "\n"
         return out
 
 
@@ -153,17 +218,17 @@ def main_cli():
       3. Batch processing
       4. Listing
     '''
-    controller = ShowController()
+    controller = Controller()
     controller.setup()
 
     repo = Repo(controller.mediapath)
     repo.scan()
     repo.load(FilenameMatches.All_Media)
-    movies = repo.get_movies()
-    movie_report = MovieShowReport()
-    movie_report.set_movies(movies)
-    movie_report.params_from_controller(controller)
-    print(movie_report.report())
+    albums = repo.get_albums()
+    album_list = AlbumList()
+    album_list.set_albums(albums)
+    album_list.params_from_controller(controller)
+    print(album_list.report())
 
 
 if __name__ == '__main__':